+++ docs/tr2/CHANGELOG.md
@@ -12,7 +12,7 @@
 - added graphics effects to the savegame so they now persist on load (#2736)
 - added an option to control whether or not Lara responds to hitting a wall while wading (#3138)
 - added an option to fix the breakable floor descending glitch (#3152)
-- added an option to fix jumping against walls in a certain way allowing teleporting (#3153)
+- added an option to fix wall glitches, or to use TR1 wall glitch behaviour (#3153)
 - added an option to disable swing cancelling (#3150)
 - added an option to disable lean jumping (#3151)
 - added an option to disable smooth wall deflection when Lara comes to a stop at a wall, similar to TR1 (#3148)